Move Rating to student model package
