+++ src/main/java/tatracker/logic/parser/ParserUtil.java
@@ -13,12 +13,12 @@ import tatracker.commons.core.index.Index;
 import tatracker.commons.util.StringUtil;
 import tatracker.logic.parser.exceptions.ParseException;
 import tatracker.model.group.Group.GroupType;
-import tatracker.model.rating.Rating;
 import tatracker.model.session.SessionType;
 import tatracker.model.student.Email;
 import tatracker.model.student.Matric;
 import tatracker.model.student.Name;
 import tatracker.model.student.Phone;
+import tatracker.model.student.Rating;
 import tatracker.model.tag.Tag;
 
 /**
