+++ src/main/java/tatracker/model/student/Student.java
@@ -9,7 +9,6 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 
-import tatracker.model.rating.Rating;
 import tatracker.model.tag.Tag;
 
 /**
