+++ src/test/java/tatracker/model/rating/RatingTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
+import tatracker.model.student.Rating;
 import tatracker.testutil.Assert;
 
 public class RatingTest {
